remove nargs register from rtl vm
It used to be that we tried to pass the argument count in a register.
We might still do this when compiling natively, but in the VM, having an
additional piece of mutable state in the VM is an unneeded distraction
to GCC's register allocator.  Also it's unneeded now that we update the
VM's stack pointer -- and it doesn't look like removing the stack
pointer entirely will be profitable.

* libguile/vm-engine.c (FRAME_LOCALS_COUNT): New helper, replaces use of
  nargs.  Adapt callers.
  (subr-call, foreign-call, continuation-call): Remove nargs parameter.
  (compose-continuation): Rename from partial-cont-call and remove nargs
  parameter.